Exchange the data classes with ones generated from OpenAPI spec (#20)
* Exchange the data classes with ones generated from OpenAPI spec

Replaces the data classes that used to be hand-written with classes that are generated from the API server's OpenAPI spec. The generated classes are included via a Git submodule, so the README/documentation will need updating.

Furthermore, the currently used submodule version does not work out of the box, as I had to fix a missing property in the generator config. This will need to be fixed in the OpenAPI project and the submodule updated.

Lastly, there is a bit of cleaning to do, as the project still contains the old, manually written classes. Adding a script to generate the OpenAPI classes might also be useful.

Signed-off-by: Sebastian Kaupe <kaupe@gonicus.de>

* Update devcontainer files

Signed-off-by: Sebastian Kaupe <kaupe@gonicus.de>

* fix query URL format to avoid parser error in API server

Signed-off-by: Sebastian Kaupe <kaupe@gonicus.de>

* update config for use with local Kind env

Signed-off-by: Sebastian Kaupe <kaupe@gonicus.de>

* remove Git submodule for status server API prior
to reintroduction under the src directory

Signed-off-by: Sebastian Kaupe <kaupe@gonicus.de>

* move status-api client into src tree

Signed-off-by: Sebastian Kaupe <kaupe@gonicus.de>

* adapt config for new severity scheme

Signed-off-by: Sebastian Kaupe <kaupe@gonicus.de>

* remove testdata and update README

Signed-off-by: Sebastian Kaupe <kaupe@gonicus.de>

---------

Signed-off-by: Sebastian Kaupe <kaupe@gonicus.de>